sigh optimizations & fixes
