+++ sigh/lib/sigh.rb
@@ -1,19 +1,4 @@
-require 'fastlane_core'
-require 'sigh/resign'
-require 'sigh/manager'
-require 'sigh/options'
-require 'sigh/local_manage'
-
-module Sigh
-  # Use this to just setup the configuration attribute and set it later somewhere else
-  class << self
-    attr_accessor :config
-  end
-
-  Helper = FastlaneCore::Helper # you gotta love Ruby: Helper.* should use the Helper class contained in FastlaneCore
-  UI = FastlaneCore::UI
-  ROOT = Pathname.new(File.expand_path('../..', __FILE__))
-
-  ENV['FASTLANE_TEAM_ID'] ||= ENV["SIGH_TEAM_ID"]
-  ENV['DELIVER_USER'] ||= ENV["SIGH_USERNAME"]
-end
+require_relative 'sigh/resign'
+require_relative 'sigh/manager'
+require_relative 'sigh/options'
+require_relative 'sigh/local_manage'
