+++ sigh/lib/sigh/options.rb
@@ -1,5 +1,5 @@
-require 'fastlane_core'
-require 'credentials_manager'
+require 'fastlane_core/configuration/configuration'
+require 'credentials_manager/appfile_config'
 
 module Sigh
   class Options
